Remove BuyBottomBar and replace it by BottomBar
